LayoutTests/imported/w3c/web-platform-tests/cookie-store/cookieStore_delete_basic.https.any-expected.txt


PASS cookieStore.delete return type is Promise<void>
